implemented the pall opcode
