+++ monty.h
@@ -1,10 +1,13 @@
-#ifndef MONTY_H
-#define MONTY_H
+#ifndef __MONTY_H__
+#define __MONTY_H__
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <string.h>
-#include <errno.h>
+#include <unistd.h>
+
+#define DELIMS " \n\t"
+
+extern char **all_op_tokens;
 
 /**
  * struct stack_s - doubly linked list representation of a stack (or queue)
@@ -13,13 +16,13 @@
  * @next: points to the next element of the stack (or queue)
  *
  * Description: doubly linked list node structure
- * for stack, queues, LIFO, FIFO
+ * for stack, queues, LIFO, FIFO Holberton project
  */
 typedef struct stack_s
 {
-    int n;
-    struct stack_s *prev;
-    struct stack_s *next;
+	int n;
+	struct stack_s *prev;
+	struct stack_s *next;
 } stack_t;
 
 /**
@@ -28,23 +31,33 @@ typedef struct stack_s
  * @f: function to handle the opcode
  *
  * Description: opcode and its function
- * for stack, queues, LIFO, FIFO
+ * for stack, queues, LIFO, FIFO Holberton project
  */
 typedef struct instruction_s
 {
-    char *opcode;
-    void (*f)(stack_t **stack, unsigned int line_number);
+	char *opcode;
+	void (*f)(stack_t **stack, unsigned int line_number);
 } instruction_t;
 
-extern char **all_Op_Tokens;
-extern int execFunc_exitStatus;
+void free_stack(stack_t **stack);
+int init_stack(stack_t **stack);
+int check_mode(stack_t *stack);
 
-int run_monty(FILE *filepathName);
-void free_Opcode(int token_Count);
-int execute_Opcode(stack_t **stack, unsigned int line_number);
+void free_op_tokens(void);
+unsigned int token_arr_len(void);
+int run_monty(FILE *file_pathname);
+void set_op_tok_error(int exit_status);
+
 void push_to_stack(stack_t **stack, unsigned int line_number);
 void print_all_stack(stack_t **stack, unsigned int line_number);
-int is_valid_int(char *value);
-void free_stack(stack_t **stack);
+
+char **strtow(char *str, char *delims);
+char *change_int_str(int n);
+
+int usage_error(void);
+int malloc_error(void);
+int file_open_error(char *filename);
+int unknown_op_error(char *opcode, unsigned int line_number);
+int no_int_error(unsigned int line_number);
 
 #endif
