--- my_monty/my_monty/test
@@ -1,16 +1,0 @@
-#include <stdio.h>
-
-int main(void)
-{
-	int status;
-	char *lineptr;
-	size_t n = 0;
-
-	if (getline(&lineptr, &n, stdin) == -1)
-		printf("didnt work");
-	else
-		printf("perfect");
-
-	return (0);
-
-}
