+++ old_monty/monty.h
@@ -1,0 +1,50 @@
+#ifndef MONTY_H
+#define MONTY_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <errno.h>
+
+/**
+ * struct stack_s - doubly linked list representation of a stack (or queue)
+ * @n: integer
+ * @prev: points to the previous element of the stack (or queue)
+ * @next: points to the next element of the stack (or queue)
+ *
+ * Description: doubly linked list node structure
+ * for stack, queues, LIFO, FIFO
+ */
+typedef struct stack_s
+{
+    int n;
+    struct stack_s *prev;
+    struct stack_s *next;
+} stack_t;
+
+/**
+ * struct instruction_s - opcode and its function
+ * @opcode: the opcode
+ * @f: function to handle the opcode
+ *
+ * Description: opcode and its function
+ * for stack, queues, LIFO, FIFO
+ */
+typedef struct instruction_s
+{
+    char *opcode;
+    void (*f)(stack_t **stack, unsigned int line_number);
+} instruction_t;
+
+extern char **all_Op_Tokens;
+extern int execFunc_exitStatus;
+
+int run_monty(FILE *filepathName);
+void free_Opcode(int token_Count);
+int execute_Opcode(stack_t **stack, unsigned int line_number);
+void push_to_stack(stack_t **stack, unsigned int line_number);
+void print_all_stack(stack_t **stack, unsigned int line_number);
+int is_valid_int(char *value);
+void free_stack(stack_t **stack);
+
+#endif
