+++ stack_helper_functions.c
